Question 6
First I check if a is null. If a != null then I proceed to check if
a is empty using the String.isEmpty() function. If the string is not
null and is not empty then it has to contain text. Return null if null
and bypass is empty and contains text options.