Question 15
I created a variable to store the sum as an Integer. Then I used a for
loop to iterate through the List as long as i was less than the size of
the List. In the loop I first checked if the value at that index in the
List was < 0. If this was true I used continue; to skip that iteration
of the loop. If it was false then I added that value to the sum. Lastly,
I returned the sum.