Question 16
I instantiated a found variable and NOT_FOUND variable to start. Then I
used a for loop with an if else statement to check each index in the
List. If found I break out of the loop, if not found then continue.
Then I run a quick if else statement to determine if the found variable
was populated or if NOT_FOUND should be returned.